+++ src/main/java/com/goldenfield192/irpatches/accessor/IRailSettingsMutableAccessor.java
@@ -1,18 +1,4 @@
 package com.goldenfield192.irpatches.accessor;
 
 public interface IRailSettingsMutableAccessor extends IRailSettingsAccessor {
-    //Ctrl2
-    void setNearEnd(float degree);
-
-    //Ctrl1
-    void setFarEnd(float degree);
-
-    //Bumpiness
-    void setBumpiness(float factor);
-
-    float getNearEndTilt();
-
-    float getFarEndTilt();
-
-    float getBumpiness();
 }